Folded util::contiguous_sequence_wrapper<T> into coretmpl.h, and fixing the reverse_iterators
